Changed spm library to run on floats rather than doubles

+++ GizmoPlatformio/lib/SPMController/SPMController.cpp
@@ -8,43 +8,43 @@ cMotor({sz_angle_sin,0,-cz_angle_cos})
 {}
 
 
-void SPMController::begin(double * a_motor_ptr, double * b_motor_ptr){
+void SPMController::begin(float * a_motor_ptr, float * b_motor_ptr){
   this-> a_motor_ptr = a_motor_ptr;
   this-> b_motor_ptr = b_motor_ptr;
 }
 
-void SPMController::calculate_motors(double phi, double theta){
+void SPMController::calculate_motors(float phi, float theta){
   //Find direction vector given angle
-  vector <double> driver_arm = get_direction_vector(phi,theta);
+  vector <float> driver_arm = get_direction_vector(phi,theta);
   //print_vector(driver_arm, "DRIVER ARM");
   //Find joint c (attached to static motor)
-  vector <double> cJoint = cross_product(cMotor, driver_arm);
+  vector <float> cJoint = cross_product(cMotor, driver_arm);
   //print_vector(cJoint, "JOINT C");
 
   //Get cross part of quaternion rotation
-  vector <double> rotationCross = scaxvec(cross_product(driver_arm, cJoint),sqrt(3)/2);
+  vector <float> rotationCross = scaxvec(cross_product(driver_arm, cJoint),sqrt(3)/2);
   //print_vector(rotationCross, "ROTATION CROSS");
 
   //Get constant part of vector rotation
-  vector <double> rotConst = scaxvec(cJoint, -0.5);
+  vector <float> rotConst = scaxvec(cJoint, -0.5);
   //print_vector(rotConst, "Rotation constant");
 
   //Get A joint position
-  vector <double> aJoint = sub_vectors(rotConst, rotationCross);
+  vector <float> aJoint = sub_vectors(rotConst, rotationCross);
   //print_vector(aJoint, "A JOINT");
 
   //Get B joint position
-  vector <double> bJoint = add_vectors(rotConst, rotationCross);
+  vector <float> bJoint = add_vectors(rotConst, rotationCross);
   //print_vector(bJoint, "B JOINT");
 
   //Get A motor position
-  double aMotor = get_motor_angle(aJoint);
+  float aMotor = get_motor_angle(aJoint);
   //Serial.println("MOTOR A ANGLE");
   //Serial.println(aMotor);
   //Serial.println();
 
   //Get B motor position
-  double bMotor = get_motor_angle(bJoint);
+  float bMotor = get_motor_angle(bJoint);
   //Serial.println("MOTOR B ANGLE");
   //Serial.println(bMotor);
   //Serial.println();
@@ -57,11 +57,11 @@ void SPMController::calculate_motors(double phi, double theta){
 
 
 //Function to get the unit vector of the driver arm with a given angle, where theta is angle around the z axis and phi is the angle from the z axis
-vector <double> SPMController::get_direction_vector(double phi, double theta)
+vector <float> SPMController::get_direction_vector(float phi, float theta)
 {
-  vector <double> driver_arm(3);
-    double rphi_rads = phi*pi/180;
-    double rtheta_rads = theta*pi/180;
+  vector <float> driver_arm(3);
+    float rphi_rads = phi*pi/180;
+    float rtheta_rads = theta*pi/180;
 
     driver_arm[0] = sin(rphi_rads)*cos(rtheta_rads);
     driver_arm[1] = sin(rphi_rads)*sin(rtheta_rads);
@@ -70,8 +70,8 @@ vector <double> SPMController::get_direction_vector(double phi, double theta)
 }
 
 //Function to get the angle of the joint vector in the xy plane
-double SPMController::get_joint_angle(double x,double y){
-  double angle;
+float SPMController::get_joint_angle(float x,float y){
+  float angle;
   //check the quadrant of the angle to increment it the correct amount
   if (x<0){
     angle = atan(y/x) + pi;
@@ -93,18 +93,18 @@ double SPMController::get_joint_angle(double x,double y){
 }
 
 //Function to solve for the motor angle from the x axis, knowing the vector of the joint
-double SPMController::get_motor_angle(vector <double> joint){
+float SPMController::get_motor_angle(vector <float> joint){
   //Define parameters of equation mcosx + nsinx = p where x is the desired angle
-  double m = joint[0]*sz_angle_sin;
-  double n = joint[1]*sz_angle_sin;
-  double p = joint[2]*cz_angle_cos;
+  float m = joint[0]*sz_angle_sin;
+  float n = joint[1]*sz_angle_sin;
+  float p = joint[2]*cz_angle_cos;
   //Define gamma, such that cos(x-gamma) = p/sqrt(n^2+m^2)
-  double gamma = get_joint_angle(joint[0],joint[1]);
+  float gamma = get_joint_angle(joint[0],joint[1]);
 
   //Define q, where q is p/sqrt(m^2+n^2)
-  double q = p/sqrt(m*m+n*n);
+  float q = p/sqrt(m*m+n*n);
   //Generate angle
-  double motor_angle = pi-acos(q)+gamma;
+  float motor_angle = pi-acos(q)+gamma;
   if (motor_angle > 2*pi){
     motor_angle -= 2*pi;
   }
@@ -116,8 +116,8 @@ double SPMController::get_motor_angle(vector <double> joint){
 
 
 //Function to calculate the cross product of 2 vectors
-vector <double> SPMController::cross_product(vector<double> a, vector <double> b){
-  vector <double> c(3); // Initialize vector with 3 elements
+vector <float> SPMController::cross_product(vector<float> a, vector <float> b){
+  vector <float> c(3); // Initialize vector with 3 elements
   c[0] = a[1]*b[2]-a[2]*b[1];
   c[1] = a[2]*b[0]-a[0]*b[2];
   c[2] = a[0]*b[1]-a[1]*b[0];
@@ -126,8 +126,8 @@ vector <double> SPMController::cross_product(vector<double> a, vector <double> b
 
 
 //Function to multiply a vector by a scalar
-vector <double> SPMController::scaxvec(vector <double> vec, double sca){
-  vector <double> new_vec(3);
+vector <float> SPMController::scaxvec(vector <float> vec, float sca){
+  vector <float> new_vec(3);
   int i;
   for (i=0; i<vec.size();i++){
     new_vec[i] = vec[i]*sca;
@@ -137,8 +137,8 @@ vector <double> SPMController::scaxvec(vector <double> vec, double sca){
 
 
 //Function to add 2 vectors
-vector <double> SPMController::add_vectors(vector <double> a, vector <double> b){
-  vector <double> new_vec(3);
+vector <float> SPMController::add_vectors(vector <float> a, vector <float> b){
+  vector <float> new_vec(3);
   int i;
   for (i=0; i<a.size();i++){
     new_vec[i] = a[i]+b[i];
@@ -148,8 +148,8 @@ vector <double> SPMController::add_vectors(vector <double> a, vector <double> b)
 
 
 //Function to subtract a vector from another
-vector <double> SPMController::sub_vectors(vector <double> a, vector <double> b){
-  vector <double> new_vec(3);
+vector <float> SPMController::sub_vectors(vector <float> a, vector <float> b){
+  vector <float> new_vec(3);
   int i;
   for (i=0; i<a.size();i++){
     new_vec[i] = a[i]-b[i];
@@ -158,7 +158,7 @@ vector <double> SPMController::sub_vectors(vector <double> a, vector <double> b)
 }
 
 //Function to serial print a vector
-void SPMController::print_vector(vector <double> to_print, String title ){
+void SPMController::print_vector(vector <float> to_print, String title ){
   int i;
   Serial.println(title);
   for (i=0; i<to_print.size(); i++){
